DEV Refactoring project
