+++ web_api.py
@@ -1,7 +1,8 @@
 from http.server import HTTPServer, BaseHTTPRequestHandler
-import main
+import searcher
 import os
 
+CSV_FILE_PATH = os.path.abspath(".\\tmp\\separated csv\\")
 
 class ServiceHandler(BaseHTTPRequestHandler):
     def do_GET(self):
@@ -28,27 +29,16 @@ class ServiceHandler(BaseHTTPRequestHandler):
                 sku = arg.split('=')[1]
             elif "rank" in arg:
                 rank = arg.split('=')[1]
-        if isinstance(sku,type(None)):
+        if isinstance(sku, type(None)):
             self.end_headers()
             self.wfile.write(bytes("Sorry, but sku is obligatory, rank is optional", 'utf-8'))
             return
-        result = main.find_some_row(sku, rank)
+        result = searcher.find_some_row(sku, CSV_FILE_PATH, rank)
         self.end_headers()
         self.wfile.write(bytes(f"{result}", 'utf-8'))
         return
 
 
 if __name__ == '__main__':
-    if os.path.isdir(".\\tmp") is False:
-        os.mkdir('.\\tmp')
-    if os.path.isdir(".\\tmp\\test") is False:
-        os.mkdir(".\\tmp\\test")
-
-    new_scv_name = 'new_recommends.csv'
-    if os.path.isfile(f".\\tmp\\{new_scv_name}") is False:
-        main.create_dict_from_csv()
-
-    main.csv_separator()
-
     server = HTTPServer(('127.0.0.1', 5000), ServiceHandler)
     server.serve_forever()
